Move storage tests out of Bank
