+++ programs/native/storage/Cargo.toml
@@ -16,6 +16,9 @@ serde_derive = "1.0.88"
 solana-logger = { path = "../../../logger", version = "0.12.0" }
 solana-sdk = { path = "../../../sdk", version = "0.12.0" }
 
+[dev-dependencies]
+solana-runtime = { path = "../../../runtime", version = "0.12.0" }
+
 [lib]
 name = "solana_storage_program"
 crate-type = ["cdylib"]
